feat: provide a default init function that matches the go sdk

+++ src/params.rs
@@ -44,7 +44,7 @@ pub struct RunParameters {
     pub test_capture_profiles: String, // TEST_CAPTURE_PROFILES:
 
     #[clap(env)]
-    pub test_group_instance_count: usize, // TEST_GROUP_INSTANCE_COUNT: 1
+    pub test_group_instance_count: u64, // TEST_GROUP_INSTANCE_COUNT: 1 // TODO: should match the test_instance_count type (u64)
     #[clap(env)]
     pub test_group_id: String, // TEST_GROUP_ID: single
